Implemented solutions for 5 Practice-Perfect problems along with test cases
