+++ anneshwa/01_Reverse_a_String/README.md
@@ -1,0 +1,12 @@
+Problem 1: Reverse a String
+Implement a function that reverses a string.
+
+Example:
+Input: "hello"
+Output: "olleh"
+
+Unit Tests:
+1. Test with a non-empty string
+2. Test with an empty string
+3. Test with a string containing spaces
+4. Test with a palindrome
