+++ anneshwa/01_Reverse_a_String/go.mod
@@ -1,0 +1,10 @@
+module 01_reverse_a_string
+
+go 1.22.5
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/testify v1.9.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
