+++ anneshwa/01_Reverse_a_String/main_test.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"testing"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestReverseString(t *testing.T) {
+	assert.Equal(t, "olleh", ReverseString("hello"))
+	assert.Equal(t, "", ReverseString(""))
+	assert.Equal(t, "  dlrow olleh  ", ReverseString("  hello world  "))
+	assert.Equal(t, "racecar", ReverseString("racecar"))
+}
