+++ anneshwa/02_valid_anagram/README.md
@@ -1,0 +1,12 @@
+Problem 2: Valid Anagram
+Given two strings s and t, return true if t is an anagram of s, and false otherwise.
+
+Example:
+Input: s = "anagram", t = "nagaram"
+Output: true
+
+Unit Tests:
+1. Test with valid anagrams
+2. Test with strings of different lengths
+3. Test with strings containing the same letters but different frequencies
+4. Test with empty strings
