+++ anneshwa/04_remove_duplicates/README.md
@@ -1,0 +1,12 @@
+Problem 4: Remove Duplicates from Sorted Array
+Given a sorted array nums, remove the duplicates in-place such that each element appears only once and returns the new length.
+
+Example:
+Input: nums = [1,1,2]
+Output: 2, nums = [1,2,_]
+
+Unit Tests:
+1. Test with an array containing duplicates
+2. Test with an array with no duplicates
+3. Test with an empty array
+4. Test with an array containing all identical elements
