+++ anneshwa/05_longest_common_prefix/README.md
@@ -1,0 +1,12 @@
+Problem 5: Longest Common Prefix
+Write a function to find the longest common prefix string amongst an array of strings.
+
+Example:
+Input: strs = ["flower","flow","flight"]
+Output: "fl"
+
+Unit Tests:
+1. Test with an array of strings with a common prefix
+2. Test with an array of strings with no common prefix
+3. Test with an array containing an empty string
+4. Test with an array containing only one string
