--- anneshwa/tmp/build-errors.log
@@ -1,1 +1,0 @@
-exit status 1exit status 1exit status 1exit status 1exit status 1exit status 1exit status 1
